refactor(ui): fix small issues around navigation (#1822)
* Simplify horizontal menu navigation

* Move the back button out of the design system
  It is a fragment of the Project screen.

* Split Header component into screen fragments
  It was doing too much with too much branching.

* Remove slot wrappers
