--- ui/DesignSystem/Component/Header/index.ts
@@ -1,7 +1,0 @@
-import Back from "./Back.svelte";
-import Large from "./Large.svelte";
-
-export default {
-  Back,
-  Large,
-};
