+++ ui/Screen/Profile.svelte
@@ -3,16 +3,20 @@
   import Router from "svelte-spa-router";
 
   import { isExperimental } from "../src/config";
+  import * as modal from "../src/modal";
   import * as path from "../src/path";
   import type { UnsealedSession } from "../src/session";
   import { settings } from "../src/session";
 
-  import { Icon } from "../DesignSystem/Primitive";
+  import { Button, Icon } from "../DesignSystem/Primitive";
 
   import ActionBar from "../DesignSystem/Component/ActionBar.svelte";
-  import Header from "../DesignSystem/Component/Header/Large.svelte";
+  import Header from "../DesignSystem/Component/Header.svelte";
   import HorizontalMenu from "../DesignSystem/Component/HorizontalMenu.svelte";
   import SidebarLayout from "../DesignSystem/Component/SidebarLayout.svelte";
+  import ProfileHeader from "./Profile/ProfileHeader.svelte";
+
+  import ModalNewProject from "../Modal/NewProject.svelte";
 
   import Following from "./Profile/Following.svelte";
   import Projects from "./Profile/Projects.svelte";
@@ -24,14 +28,6 @@
     "/profile/projects": Projects,
     "/profile/funding": Funding,
     "*": NotFound,
-  };
-
-  import ProjectsMenu from "./Profile/ProjectsMenu.svelte";
-
-  const menuRoutes = {
-    "/profile/projects": ProjectsMenu,
-    "/profile/tracking": ProjectsMenu,
-    "*": ProjectsMenu,
   };
 
   const topbarMenuItems = [
@@ -59,19 +55,27 @@
 </script>
 
 <SidebarLayout style="margin-top: 0;" dataCy="profile-screen">
-  <Header
-    avatarFallback={session.identity.avatarFallback}
-    name={session.identity.metadata.handle}
-    peerId={session.identity.peerId}>
-    <div slot="top">
-      <Router routes={menuRoutes} />
-    </div>
+  <Header>
+    <ProfileHeader
+      slot="left"
+      avatarFallback={session.identity.avatarFallback}
+      name={session.identity.metadata.handle}
+      peerId={session.identity.peerId} />
+
+    <Button
+      slot="right"
+      dataCy="new-project-button"
+      variant="outline"
+      icon={Icon.Plus}
+      on:click={() => {
+        modal.toggle(ModalNewProject);
+      }}>
+      New project
+    </Button>
   </Header>
 
   <ActionBar>
-    <div slot="left">
-      <HorizontalMenu items={topbarMenuItems} />
-    </div>
+    <HorizontalMenu slot="left" items={topbarMenuItems} />
   </ActionBar>
 
   <Router routes={screenRoutes} />
